+++ vscode/src/chat/ChatViewProvider.ts
@@ -64,6 +64,7 @@ export type Config = Pick<
     | 'useContext'
     | 'experimentalChatPredictions'
     | 'experimentalGuardrails'
+    | 'experimentalCustomRecipes'
     | 'pluginsEnabled'
     | 'pluginsConfig'
     | 'pluginsDebugEnabled'
@@ -127,7 +128,6 @@ export class ChatViewProvider implements vscode.WebviewViewProvider, vscode.Disp
         this.disposables.push(this.configurationChangeEvent)
 
         this.currentWorkspaceRoot = ''
-        this.customRecipesInit()
 
         // listen for vscode active editor change event
         this.disposables.push(
@@ -192,6 +192,7 @@ export class ChatViewProvider implements vscode.WebviewViewProvider, vscode.Disp
         if (authStatus.endpoint) {
             this.config.serverEndpoint = authStatus.endpoint
         }
+        void this.sendMyPrompts()
         this.configurationChangeEvent.fire()
     }
 
@@ -298,7 +299,7 @@ export class ChatViewProvider implements vscode.WebviewViewProvider, vscode.Disp
                 void this.openExternalLinks(message.value)
                 break
             case 'my-prompt':
-                await this.executeMyPrompt(message.title)
+                await this.executeCustomRecipe(message.title)
                 break
             case 'openFile': {
                 const rootPath = this.editor.getWorkspaceRootPath()
@@ -733,7 +734,10 @@ export class ChatViewProvider implements vscode.WebviewViewProvider, vscode.Disp
         })
     }
 
-    private async executeMyPrompt(title: string): Promise<void> {
+    public async executeCustomRecipe(title: string): Promise<void> {
+        if (!this.config.experimentalCustomRecipes) {
+            return
+        }
         // Send prompt names to webview to display as recipe options
         if (!title || title === 'get') {
             await this.sendMyPrompts()
@@ -757,9 +761,9 @@ export class ChatViewProvider implements vscode.WebviewViewProvider, vscode.Disp
         }
         // Get prompt details from controller by title then execute prompt's command
         const prompt = this.editor.controllers.prompt.find(title)
-        this.editor.controllers.prompt.getCommandOutput()
+        await this.editor.controllers.prompt.get('command')
         if (!prompt) {
-            debug('executeMyPrompt:noPrompt', title)
+            debug('executeCustomRecipe:noPrompt', title)
             return
         }
         if (!prompt.startsWith('/')) {
@@ -768,26 +772,24 @@ export class ChatViewProvider implements vscode.WebviewViewProvider, vscode.Disp
         await this.executeChatCommands(prompt, 'my-prompt')
     }
 
-    private customRecipesInit(): void {
-        // listen for file change event for JSON files used for building Custom Recipes
-        const myWorkspacePromptsWatcher = this.editor.controllers?.prompt?.wsFileWatcher
-        const myUserPromptsWatcher = this.editor.controllers?.prompt?.userFileWatcher
-        myWorkspacePromptsWatcher?.onDidChange(() => this.sendMyPrompts())
-        myWorkspacePromptsWatcher?.onDidDelete(() => this.sendMyPrompts())
-        myUserPromptsWatcher?.onDidChange(() => this.sendMyPrompts())
-        myUserPromptsWatcher?.onDidDelete(() => this.sendMyPrompts())
-    }
-
     /**
      * Send custom recipe names to webview
      */
     private async sendMyPrompts(): Promise<void> {
-        await this.editor.controllers.prompt.refresh()
-        const prompts = this.editor.controllers.prompt.getPromptList()
-        void this.webview?.postMessage({
-            type: 'my-prompts',
-            prompts,
-        })
+        const send = async (): Promise<void> => {
+            await this.editor.controllers.prompt.refresh()
+            const prompts = this.editor.controllers.prompt.getPromptList()
+            void this.webview?.postMessage({
+                type: 'my-prompts',
+                prompts,
+                isEnabled: this.config.experimentalCustomRecipes,
+            })
+        }
+        const init = (): void => {
+            this.editor.controllers.prompt.setMessager(send)
+        }
+        init()
+        await send()
     }
 
     private async saveTranscriptToChatHistory(): Promise<void> {