+++ vscode/src/main.ts
@@ -1,5 +1,3 @@
-import { debug } from 'console'
-
 import * as vscode from 'vscode'
 
 import { RecipeID } from '@sourcegraph/cody-shared/src/chat/recipes/recipe'
@@ -87,18 +85,18 @@ const register = async (
 }> => {
     const disposables: vscode.Disposable[] = []
 
+    // Controller for Inline Chat
     await createOrUpdateEventLogger(initialConfig, localStorage)
     // Controller for inline Chat
     const commentController = new InlineController(context.extensionPath)
-
+    // Controller for Non-Stop Cody
     const fixup = new FixupController()
     disposables.push(fixup)
     if (TestSupport.instance) {
         TestSupport.instance.fixupController.set(fixup)
     }
-
-    const prompt = new MyPromptController(debug, context, initialConfig.serverEndpoint)
-
+    // Controller for Custom Recipes
+    const prompt = new MyPromptController(context, initialConfig.experimentalCustomRecipes)
     const controllers = { inline: commentController, fixups: fixup, prompt }
 
     const editor = new VSCodeEditor(controllers)
@@ -216,6 +214,8 @@ const register = async (
             await chatProvider.clearHistory()
         }),
         // Recipes
+        vscode.commands.registerCommand('cody.customRecipes.exec', title => chatProvider.executeCustomRecipe(title)),
+        vscode.commands.registerCommand('cody.customRecipes.list', () => prompt.quickRecipe()),
         vscode.commands.registerCommand('cody.recipe.explain-code', () => executeRecipe('explain-code-detailed')),
         vscode.commands.registerCommand('cody.recipe.explain-code-high-level', () =>
             executeRecipe('explain-code-high-level')